% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsc_make_grid_over.R
\name{vsc_make_grid_over}
\alias{vsc_make_grid_over}
\title{Build a polygon grid over the extent of an object}
\usage{
vsc_make_grid_over(x, cellsize = NULL, n = c(50, 50), square = TRUE)
}
\arguments{
\item{x}{A \code{SpatRaster}, \code{stars}, or \code{sf} object providing the extent/CRS.}

\item{cellsize}{Numeric length (map units) of grid cells. If \code{NULL}, use \code{n}.}

\item{n}{Integer vector of length 2: number of cells in x and y (used when
\code{cellsize} is \code{NULL}). Default \code{c(50, 50)}.}

\item{square}{Logical; if \code{FALSE}, allow non-square cells if needed.}
}
\value{
An \code{sf} polygon grid with an \code{id} column.
}
\description{
Creates an \code{sf} polygon grid over the bounding box of \code{x}. You can define
the grid by a target cellsize or by number of cells per side.
}
\examples{
\dontrun{
cl_bel = vsc_build_worldclim("BEL", month = 5)
grd = vsc_make_grid_over(cl_bel$predictors, n = c(50, 50))
}
}
